fix: float32 doesnt cumsum correctly, but float64 and unint32 do

+++ falcon-vis/src/core/falconArray/arrayTypes.ts
@@ -20,8 +20,8 @@ export type TypedArrayConstructor =
   | Float32ArrayConstructor
   | Float64ArrayConstructor;
 
-export const CumulativeCountsArray = Float64Array;
+export const CumulativeCountsArray = Uint32Array;
 export const CountsArray = Uint32Array;
 
-export type CumulativeCountsArrayType = Float64Array;
+export type CumulativeCountsArrayType = Uint32Array;
 export type CountsArrayType = Uint32Array;
